Upgrades sbt to 1.3.9, scala to 2.12.11, genjavadoc to 0.16.
sbt, depending on the terminal you are displaying it in, will issue ANSI
colorization escapes. You may need to add the color=false option to your
environment: e.g., in my .bash_aliases I have

export SBT_OPTS=" -Dsbt.color=false "

Had to fix a number of deprecations.
junit.framework.Assert is deprecated. Changed to org.junit.Assert.

Assert.assertEquals for float/double now requires a delta argument. Supplied
1e-15 where needed.

Since scala 2.12.7, shifting of values by quantities expressed as Long is deprectated. Had to insert ".toInt" calls a number of places to silence these.

Changes to main.yml also fix DAFFODIL-2111 which was javadoc-related errors due
to java 11 occurring on the automated build.

DAFFODIL-2313, DAFFODIL-2111

+++ daffodil-lib/src/main/scala/passera/unsigned/SmallUInt.scala
@@ -181,19 +181,19 @@ trait SmallUInt[U <: Unsigned[U, UInt, Int]] extends Any with Unsigned[U, UInt, 
   def unary_~ = UInt(~intValue)
 
   def <<(x: Int)(implicit d: DummyImplicit) = UInt(intValue << x)
-  def <<(x: Long)(implicit d: DummyImplicit) = UInt(intValue << x)
+  def <<(x: Long)(implicit d: DummyImplicit) = UInt(intValue << x.toInt)
   def <<(x: UInt) = UInt(intValue << (x.toInt & 0x1f))
-  def <<(x: ULong) = UInt(intValue << (x.toLong & 0x1f))
+  def <<(x: ULong) = UInt(intValue << (x.toLong & 0x1f).toInt)
 
   def >>(x: Int)(implicit d: DummyImplicit) = UInt(intValue >>> x)
-  def >>(x: Long)(implicit d: DummyImplicit) = UInt(intValue >>> x)
+  def >>(x: Long)(implicit d: DummyImplicit) = UInt(intValue >>> x.toInt)
   def >>(x: UInt) = UInt(intValue >>> (x.toInt & 0x1f))
-  def >>(x: ULong) = UInt(intValue >>> (x.toLong & 0x1f))
+  def >>(x: ULong) = UInt(intValue >>> (x.toLong & 0x1f).toInt)
 
   def >>>(x: Int)(implicit d: DummyImplicit) = UInt(intValue >>> x)
-  def >>>(x: Long)(implicit d: DummyImplicit) = UInt(intValue >>> x)
+  def >>>(x: Long)(implicit d: DummyImplicit) = UInt(intValue >>> x.toInt)
   def >>>(x: UInt) = UInt(intValue >>> (x.toInt & 0x1f))
-  def >>>(x: ULong) = UInt(intValue >>> (x.toLong & 0x1f))
+  def >>>(x: ULong) = UInt(intValue >>> (x.toLong & 0x1f).toInt)
 
   override def toString = (intValue & 0xffffffffL).toString
